refactoring...

+++ app/src/main/java/ar/edu/itba/example/api/ui/home/PayScreen.kt
@@ -251,7 +251,7 @@ fun PayScreen(viewModel: HomeViewModel) {
                 onClick = {
                     viewModel.makePayment(
                         Payment(
-                            amount = valueToPay.toDouble(),
+                            amount = valueToPay.toFloat(),
                             description = descripcion,
                             type = tipodePago,
                             cardId = selectedCard?.id,
